Abstracted backup processing logic into the BackupProcessor interface
