+++ pkg/backup/compressor_test.go
@@ -10,9 +10,10 @@ import (
 
 func TestBackupCompressors(t *testing.T) {
 	content := "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
+	processor := NewLogicalBackupProcessor()
 	tests := []struct {
 		name            string
-		newCompressorFn func(basePath string, logger logr.Logger) BackupCompressor
+		newCompressorFn func(basePath string, processor BackupProcessor, logger logr.Logger) BackupCompressor
 		fileName        string
 	}{
 		{
@@ -40,7 +41,7 @@ func TestBackupCompressors(t *testing.T) {
 			}
 			defer os.RemoveAll(dir)
 
-			compressor := tt.newCompressorFn(dir, logger)
+			compressor := tt.newCompressorFn(dir, processor, logger)
 
 			filePath := filepath.Join(dir, tt.fileName)
 			if err := os.WriteFile(filePath, []byte(content), 0644); err != nil {
